Drop codal from linux build
