+++ libs/mixer---linux/sound.cpp
@@ -1,8 +1,11 @@
 #include "pxt.h"
 #include "SoundOutput.h"
+#include "melody.h"
 
 #include <alsa/asoundlib.h>
 #include <pthread.h>
+
+namespace music {
 
 static void alsa_check(int pos, int fn) {
     if (fn < 0) {
@@ -27,18 +30,19 @@ void *LinuxDAC::play(void *self) {
     DMESG("PCM state: %s", snd_pcm_state_name(snd_pcm_state(pcm_handle)));
 
     for (;;) {
-        auto buf = dac->src.pull();
-        auto data = (int16_t *)buf.getBytes();
-        auto len = buf.length() / 2;
-        if (!len) {
+        target_disable_irq();
+        auto hasData = dac->src.fillSamples(dac->data, sizeof(dac->data) / 2);
+        target_enable_irq();
+        auto len = (int)sizeof(dac->data) / 2;
+        if (!hasData) {
             sleep_core_us(5000);
             continue;
         }
         for (int i = 0; i < len; ++i) {
             // playing at half-volume
-            data[i] = (data[i] - 512) << 6;
+            dac->data[i] = dac->data[i] << 3;
         }
-        int frames = snd_pcm_writei(pcm_handle, data, len);
+        int frames = snd_pcm_writei(pcm_handle, dac->data, len);
         if (frames < 0)
             frames = snd_pcm_recover(pcm_handle, frames, 0);
         if (frames < 0) {
@@ -50,9 +54,10 @@ void *LinuxDAC::play(void *self) {
     return NULL;
 }
 
-LinuxDAC::LinuxDAC(DataSource &data) : src(data) {
-    data.connect(*this);
+LinuxDAC::LinuxDAC(WSynthesizer &data) : src(data) {
     pthread_t upd;
     pthread_create(&upd, NULL, LinuxDAC::play, this);
     pthread_detach(upd);
+}
+
 }
